fixed failing ut
