+++ pkg/apis/zookeeper/v1beta1/zookeepercluster_types.go
@@ -12,14 +12,11 @@ package v1beta1
 
 import (
 	"fmt"
-	"time"
 
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-type ClusterConditionType string
 
 const (
 	// DefaultZkContainerRepository is the default docker repo for the zookeeper
@@ -37,14 +34,6 @@ const (
 	// container is stopped. This gives clients time to disconnect from a
 	// specific node gracefully.
 	DefaultTerminationGracePeriod = 30
-
-	ClusterConditionPodsReady ClusterConditionType = "PodsReady"
-	ClusterConditionUpgrading                      = "Upgrading"
-	ClusterConditionError                          = "Error"
-
-	// Reasons for cluster upgrading condition
-	UpdatingZookeeperReason = "Updating Zookeeper"
-	UpgradeErrorReason      = "Upgrade Error"
 )
 
 // ZookeeperClusterSpec defines the desired state of ZookeeperCluster
@@ -132,198 +121,6 @@ func (s *ZookeeperClusterSpec) withDefaults(z *ZookeeperCluster) (changed bool) 
 		changed = true
 	}
 	return changed
-}
-
-// ZookeeperClusterStatus defines the observed state of ZookeeperCluster
-type ZookeeperClusterStatus struct {
-	// Members is the zookeeper members in the cluster
-	Members MembersStatus `json:"members"`
-
-	// Replicas is the number of number of desired replicas in the cluster
-	Replicas int32 `json:"replicas"`
-
-	// ReadyReplicas is the number of number of ready replicas in the cluster
-	ReadyReplicas int32 `json:"readyReplicas"`
-
-	// InternalClientEndpoint is the internal client IP and port
-	InternalClientEndpoint string `json:"internalClientEndpoint"`
-
-	// ExternalClientEndpoint is the internal client IP and port
-	ExternalClientEndpoint string `json:"externalClientEndpoint"`
-
-	MetaRootCreated bool `json:"metaRootCreated"`
-
-	// CurrentVersion is the current cluster version
-	CurrentVersion string `json:"currentVersion"`
-
-	TargetVersion string `json:"targetVersion"`
-
-	// Conditions list all the applied conditions
-	Conditions []ClusterCondition `json:"conditions,omitempty"`
-}
-
-// ClusterCondition shows the current condition of a Zookeeper cluster.
-// Comply with k8s API conventions
-type ClusterCondition struct {
-	// Type of Zookeeper cluster condition.
-	Type ClusterConditionType `json:"type"`
-
-	// Status of the condition, one of True, False, Unknown.
-	Status v1.ConditionStatus `json:"status"`
-
-	// The reason for the condition's last transition.
-	Reason string `json:"reason,omitempty"`
-
-	// A human readable message indicating details about the transition.
-	Message string `json:"message,omitempty"`
-
-	// The last time this condition was updated.
-	LastUpdateTime string `json:"lastUpdateTime,omitempty"`
-
-	// Last time the condition transitioned from one status to another.
-	LastTransitionTime string `json:"lastTransitionTime,omitempty"`
-}
-
-func (ps *ZookeeperClusterStatus) Init() {
-	// Initialise conditions
-	conditionTypes := []ClusterConditionType{
-		ClusterConditionPodsReady,
-		ClusterConditionUpgrading,
-		ClusterConditionError,
-	}
-	for _, conditionType := range conditionTypes {
-		if _, condition := ps.GetClusterCondition(conditionType); condition == nil {
-			c := newClusterCondition(conditionType, v1.ConditionFalse, "", "")
-			ps.setClusterCondition(*c)
-		}
-	}
-}
-
-func newClusterCondition(condType ClusterConditionType, status v1.ConditionStatus, reason, message string) *ClusterCondition {
-	return &ClusterCondition{
-		Type:               condType,
-		Status:             status,
-		Reason:             reason,
-		Message:            message,
-		LastUpdateTime:     "",
-		LastTransitionTime: "",
-	}
-}
-
-func (ps *ZookeeperClusterStatus) SetPodsReadyConditionTrue() {
-	c := newClusterCondition(ClusterConditionPodsReady, v1.ConditionTrue, "", "")
-	ps.setClusterCondition(*c)
-}
-
-func (ps *ZookeeperClusterStatus) SetPodsReadyConditionFalse() {
-	c := newClusterCondition(ClusterConditionPodsReady, v1.ConditionFalse, "", "")
-	ps.setClusterCondition(*c)
-}
-
-func (ps *ZookeeperClusterStatus) SetUpgradingConditionTrue(reason, message string) {
-	c := newClusterCondition(ClusterConditionUpgrading, v1.ConditionTrue, reason, message)
-	ps.setClusterCondition(*c)
-}
-
-func (ps *ZookeeperClusterStatus) SetUpgradingConditionFalse() {
-	c := newClusterCondition(ClusterConditionUpgrading, v1.ConditionFalse, "", "")
-	ps.setClusterCondition(*c)
-}
-
-func (ps *ZookeeperClusterStatus) SetErrorConditionTrue(reason, message string) {
-	c := newClusterCondition(ClusterConditionError, v1.ConditionTrue, reason, message)
-	ps.setClusterCondition(*c)
-}
-
-func (ps *ZookeeperClusterStatus) SetErrorConditionFalse() {
-	c := newClusterCondition(ClusterConditionError, v1.ConditionFalse, "", "")
-	ps.setClusterCondition(*c)
-}
-
-func (ps *ZookeeperClusterStatus) GetClusterCondition(t ClusterConditionType) (int, *ClusterCondition) {
-	for i, c := range ps.Conditions {
-		if t == c.Type {
-			return i, &c
-		}
-	}
-	return -1, nil
-}
-
-func (ps *ZookeeperClusterStatus) setClusterCondition(newCondition ClusterCondition) {
-	now := time.Now().Format(time.RFC3339)
-	position, existingCondition := ps.GetClusterCondition(newCondition.Type)
-
-	if existingCondition == nil {
-		ps.Conditions = append(ps.Conditions, newCondition)
-		return
-	}
-
-	if existingCondition.Status != newCondition.Status {
-		existingCondition.Status = newCondition.Status
-		existingCondition.LastTransitionTime = now
-		existingCondition.LastUpdateTime = now
-	}
-
-	if existingCondition.Reason != newCondition.Reason || existingCondition.Message != newCondition.Message {
-		existingCondition.Reason = newCondition.Reason
-		existingCondition.Message = newCondition.Message
-		existingCondition.LastUpdateTime = now
-	}
-
-	ps.Conditions[position] = *existingCondition
-}
-
-func (ps *ZookeeperClusterStatus) IsClusterInUpgradeFailedState() bool {
-	_, errorCondition := ps.GetClusterCondition(ClusterConditionError)
-	if errorCondition == nil {
-		return false
-	}
-	if errorCondition.Status == v1.ConditionTrue && errorCondition.Reason == "UpgradeFailed" {
-		return true
-	}
-	return false
-}
-
-func (ps *ZookeeperClusterStatus) IsClusterInUpgradingState() bool {
-	_, upgradeCondition := ps.GetClusterCondition(ClusterConditionUpgrading)
-	if upgradeCondition == nil {
-		return false
-	}
-	if upgradeCondition.Status == v1.ConditionTrue {
-		return true
-	}
-	return false
-}
-
-func (ps *ZookeeperClusterStatus) IsClusterInReadyState() bool {
-	_, readyCondition := ps.GetClusterCondition(ClusterConditionPodsReady)
-	if readyCondition != nil && readyCondition.Status == v1.ConditionTrue {
-		return true
-	}
-	return false
-}
-
-func (ps *ZookeeperClusterStatus) UpdateProgress(reason, updatedReplicas string) {
-	if ps.IsClusterInUpgradingState() {
-		// Set the upgrade condition reason to be UpgradingBookkeeperReason, message to be 0
-		ps.SetUpgradingConditionTrue(reason, updatedReplicas)
-	}
-}
-
-func (ps *ZookeeperClusterStatus) GetLastCondition() (lastCondition *ClusterCondition) {
-	if ps.IsClusterInUpgradingState() {
-		_, lastCondition := ps.GetClusterCondition(ClusterConditionUpgrading)
-		return lastCondition
-	}
-	// nothing to do if we are neither upgrading nor rolling back,
-	return nil
-}
-
-// MembersStatus is the status of the members of the cluster with both
-// ready and unready node membership lists
-type MembersStatus struct {
-	Ready   []string `json:"ready"`
-	Unready []string `json:"unready"`
 }
 
 // +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
